requirements.txt
discord.py==2.3.2

asyncpg==0.29.0
psycopg2-binary==2.9.9

python-dotenv==1.0.1

mercadopago==2.3.0
requests==2.32.3

flask==3.0.3
